+++ apps/betterangels/src/app/(private-screens)/add-note/Location/LocationMapModal/Map.tsx
@@ -1,11 +1,6 @@
-import {
-  MapView,
-  Marker,
-  PROVIDER_GOOGLE,
-  TMapView,
-} from '@monorepo/expo/betterangels';
 import { useApiConfig } from '@monorepo/expo/shared/clients';
 import { LocationPinIcon } from '@monorepo/expo/shared/icons';
+import { MapView, Marker, PROVIDER_GOOGLE, TMapView } from '@monorepo/maps';
 import axios from 'axios';
 import * as Location from 'expo-location';
 import { forwardRef } from 'react';
