+++ apps/betterangels/src/app/(private-screens)/add-note/Location/index.tsx
@@ -1,7 +1,7 @@
-import { MapView, Marker, PROVIDER_GOOGLE } from '@monorepo/expo/betterangels';
 import { LocationPinIcon } from '@monorepo/expo/shared/icons';
 import { Colors, Spacings } from '@monorepo/expo/shared/static';
 import { FieldCard, TextMedium } from '@monorepo/expo/shared/ui-components';
+import { MapView, Marker, PROVIDER_GOOGLE } from '@monorepo/maps';
 import { RefObject, useState } from 'react';
 import { ScrollView, StyleSheet, View } from 'react-native';
 import LocationMapModal from './LocationMapModal';
