+++ libs/expo/betterangels/src/lib/index.ts
@@ -2,7 +2,7 @@ export * from './apollo';
 export * from './errors';
 export * from './helpers';
 export * from './hooks';
-export * from './maps';
+// export * from './maps';
 export * from './navigation';
 export * from './providers';
 export * from './screenRouting';
