+++ libs/expo/betterangels/src/lib/screens/Note/NoteLocation.tsx
@@ -1,8 +1,8 @@
 import { LocationPinIcon } from '@monorepo/expo/shared/icons';
 import { Colors, Radiuses } from '@monorepo/expo/shared/static';
 import { TextBold, TextRegular } from '@monorepo/expo/shared/ui-components';
+import { MapView, Marker, PROVIDER_GOOGLE } from '@monorepo/maps';
 import { StyleSheet, View } from 'react-native';
-import { MapView, Marker, PROVIDER_GOOGLE } from '../../maps';
 import { NoteSummaryQuery } from './__generated__/NoteSummary.generated';
 
 export default function NoteLocation({
