+++ libs/expo/maps/README.md
@@ -1,0 +1,11 @@
+# maps
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build maps` to build the library.
+
+## Running unit tests
+
+Run `nx test maps` to execute the unit tests via [Jest](https://jestjs.io).
