+++ libs/expo/maps/project.json
@@ -1,0 +1,26 @@
+{
+  "name": "maps",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/expo/maps/src",
+  "projectType": "library",
+  "tags": [],
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/expo/maps",
+        "main": "libs/expo/maps/src/index.ts",
+        "tsConfig": "libs/expo/maps/tsconfig.lib.json",
+        "assets": ["libs/expo/maps/*.md"]
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
+      "options": {
+        "jestConfig": "libs/expo/maps/jest.config.ts"
+      }
+    }
+  }
+}
