+++ libs/expo/shared/ui-components/src/lib/Map/Map.tsx
@@ -1,8 +1,8 @@
+import { MapView, TMapView } from '@monorepo/maps';
 import { ReactNode, useRef, useState } from 'react';
 import { StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
 import { Region } from 'react-native-maps';
 import { defaultMapRegion } from './constants';
-import { RNMapView, TRNMapView } from './mapLib';
 import { MapLocateMeBtn } from './mapUi/MapLocateMeBtn';
 
 type TMapProps = {
@@ -14,7 +14,7 @@ type TMapProps = {
   children?: ReactNode;
 };
 
-export function MapView(props: TMapProps) {
+export function Map(props: TMapProps) {
   const {
     provider,
     initialRegion,
@@ -25,11 +25,11 @@ export function MapView(props: TMapProps) {
   } = props;
 
   const [_mapReady, setMapReady] = useState(false);
-  const mapRef = useRef<TRNMapView>(null);
+  const mapRef = useRef<TMapView>(null);
 
   return (
     <View style={[styles.wrapper, style]}>
-      <RNMapView
+      <MapView
         ref={mapRef}
         provider={provider}
         showsUserLocation={!!enableUserLocation}
@@ -43,7 +43,7 @@ export function MapView(props: TMapProps) {
         style={[styles.map, mapStyle]}
       >
         {children}
-      </RNMapView>
+      </MapView>
 
       {!!enableUserLocation && <MapLocateMeBtn mapRef={mapRef} />}
     </View>
