+++ libs/expo/shared/ui-components/src/lib/Map/index.ts
@@ -1,8 +1,7 @@
 export { ClusterMap } from './ClusterMap';
-export { MapView } from './Map';
+export { Map } from './Map';
 export * from './clustering';
 export { defaultMapRegion, regionDeltaMap } from './constants';
-export * from './mapLib';
 export * from './types';
 export { coordsToRegion } from './utils/coordsToRegion';
 export { regionToBbox } from './utils/regionToBbox';
