--- libs/expo/shared/ui-components/src/lib/Map/mapLib/map.ts
@@ -1,3 +1,0 @@
-import * as Maps from 'react-native-maps';
-
-export default Maps;
