--- libs/expo/shared/ui-components/src/lib/Map/mapLib/map.web.ts
@@ -1,3 +1,0 @@
-import * as Maps from '@teovilla/react-native-web-maps';
-
-export default Maps;
