+++ libs/expo/shared/ui-components/src/lib/Map/mapUi/MapLocateMeBtn.tsx
@@ -1,12 +1,12 @@
+import { TMapView } from '@monorepo/maps';
 import { RefObject } from 'react';
 import { StyleSheet } from 'react-native';
 import LocateMeButton from '../../LocateMeButton';
-import { TRNMapView } from '../mapLib';
 import { TMapDeltaLatLng } from '../types';
 import { goToUserLocation } from '../utils/goToUserLocation';
 
 type TProps = {
-  mapRef: RefObject<TRNMapView | null>;
+  mapRef: RefObject<TMapView | null>;
   regionDelta?: TMapDeltaLatLng;
   duration?: number;
 };
